refactor: move util package to view layer
