+++ app/src/main/java/woowacourse/movie/view/ticket/TicketActivity.kt
@@ -9,9 +9,9 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
 import woowacourse.movie.R
-import woowacourse.movie.util.ErrorMessage
 import woowacourse.movie.view.reservation.ReservationActivity
 import woowacourse.movie.view.ticket.model.Ticket
+import woowacourse.movie.view.util.ErrorMessage
 
 class TicketActivity : AppCompatActivity() {
     private var ticket: Ticket? = null
